add assertion for ansible version
Added an assertion to check for Ansible version prior to running other tasks. This is due to the Ubuntu AMI failing with the standard Ansible install (apt install ansible) as the version in apt is 2.9.6 - which fails due to a missing critical update packaged into 2.9.8. If the user runs the script with anything earlier than 2.9.8, it will fail and request that they upgrade Ansible.