Refactor 2 (#54)
* Refactor 2

* add assertion for ansible version

Added an assertion to check for Ansible version prior to running other tasks. This is due to the Ubuntu AMI failing with the standard Ansible install (apt install ansible) as the version in apt is 2.9.6 - which fails due to a missing critical update packaged into 2.9.8. If the user runs the script with anything earlier than 2.9.8, it will fail and request that they upgrade Ansible.

* Revert "add assertion for ansible version"

This reverts commit 486df1e76a57cedaa5049c9a78082cea0b7aa87b.

* Add changes from review

Co-authored-by: Michael DAmato <michaeldamato@MDD-MacBook-Pro.local>
Co-authored-by: Tucker Blue <me@tuck.codes>