Added handling for unrecognized firewalld service failure (#1)
If a node has Docker installed, we run into an error where Ansible service facts collects the firewalld service since it is referred to as a dependency but actually isn't present on the machine. This means it is technically "defined" but any attempts to interact with the service will fail. This handling allows us to fail on any errors other than the service not being found, in which case we carry on (since firewalld is very much not running at that point). Confirmed working on Red Hat, Ubuntu and SLES.